Add a publication test
